+++ modules/_control_plane/versions.tf
@@ -12,7 +12,7 @@ terraform {
   required_providers {
     hcloud = {
       source  = "hetznercloud/hcloud"
-      version = "= 1.60.1"
+      version = "= 1.65.0"
     }
   }
 }